+++ examples/react/AnimatedHeader/src/App.tsx
@@ -7,11 +7,47 @@
  * LICENSE file in the root directory of this source tree.
  */
 
-import React from 'react';
+import React, { useState } from 'react';
 import { AnimatedHeader } from '@carbon-labs/react-animated-header/es/index';
+import {
+  watsonXAnimatedLight,
+  watsonXStaticLight,
+} from '@carbon-labs/react-animated-header/assets';
+import { headerTiles, workspaceData } from './data';
 
 function App() {
-  return <AnimatedHeader />;
+  const [tiles] = useState(headerTiles);
+  const [workspaces] = useState(workspaceData);
+  const [selectedTile, setSelectedTile] = useState(tiles[0]);
+  const [selectedWorkspace, setSelectedWorkspace] = useState(workspaces[0]);
+
+  const handleWorkspaceSelect = (e: any) => {
+    setSelectedWorkspace(e.selectedItem);
+  };
+
+  const handleTileGroup = (e: any) => {
+    setSelectedTile(e.selectedItem.id);
+  };
+
+  return (
+    <AnimatedHeader
+      name="Drew"
+      description="Connect, monitor, and manage data."
+      buttonText="Manage data"
+      buttonType="tertiary"
+      buttonIcon="Launch"
+      headerDropdown={false}
+      productName="[Product name]"
+      headerAnimation={watsonXAnimatedLight}
+      headerStatic={watsonXStaticLight}
+      selectedWorkspace={selectedWorkspace}
+      setSelectedWorkspace={handleWorkspaceSelect}
+      allWorkspaces={workspaces}
+      selectedTileGroup={selectedTile}
+      setSelectedTileGroup={handleTileGroup}
+      allTiles={tiles}
+    />
+  );
 }
 
 export default App;