+++ packages/react/src/components/UIShell/components/SideNav.tsx
@@ -234,16 +234,20 @@ function SideNavRenderFunction(
           }
         );
       resetNodeTabIndices();
+    }
+  }, [prefix, internalIsTreeview]);
 
+  useEffect(() => {
+    if (sideNavRef.current) {
       const firstElement = sideNavRef?.current?.querySelector(
         'a, button'
       ) as HTMLElement;
 
-      if (firstElement) {
+      if (firstElement && (navType == SIDE_NAV_TYPE.PANEL || expanded)) {
         firstElement.tabIndex = 0;
       }
     }
-  }, [prefix, internalIsTreeview]);
+  }, [expanded]);
 
   /**
    * Returns the parent SideNavMenu, if node is actually inside one.
@@ -254,7 +258,9 @@ function SideNavRenderFunction(
     const parentNode = (node as HTMLElement).parentElement?.closest(
       `.${prefix}--side-nav__item`
     );
-    if (parentNode) return parentNode;
+    if (parentNode) {
+      return parentNode;
+    }
     return node;
   }
 
@@ -502,12 +508,21 @@ function SideNavRenderFunction(
     }
   };
 
+  const SideNavToggleButton = (
+    <SideNavToggle
+      className={!expandedState ? `${prefix}--side-nav__toggle--collapsed` : ''}
+      renderIcon={expandedState ? SidePanelClose : SidePanelOpen}
+      onClick={() => setExpandedState(!expandedState)}>
+      {sideNavToggleText}
+    </SideNavToggle>
+  );
+
   return (
     <SideNavContext.Provider
       value={{
+        expanded,
         isRail,
         navType,
-        expanded: expanded,
         isTreeview: internalIsTreeview,
         setIsTreeview,
       }}>
@@ -529,13 +544,14 @@ function SideNavRenderFunction(
         {...eventHandlers}
         {...other}>
         {childrenToRender}
-        {navType === SIDE_NAV_TYPE.PANEL && (
-          <SideNavToggle
-            renderIcon={expandedState ? SidePanelClose : SidePanelOpen}
-            onClick={() => setExpandedState(!expandedState)}>
-            {sideNavToggleText}
-          </SideNavToggle>
-        )}
+        {navType === SIDE_NAV_TYPE.PANEL &&
+          (expandedState ? (
+            SideNavToggleButton
+          ) : (
+            <div className={`${prefix}--side-nav__toggle-container`}>
+              {SideNavToggleButton}
+            </div>
+          ))}
       </nav>
     </SideNavContext.Provider>
   );