+++ packages/react/src/components/UIShell/index.ts
@@ -8,6 +8,10 @@
  */
 export { SideNav, SIDE_NAV_TYPE } from './components/SideNav.js';
 export { SideNavItems } from './components/SideNavItems.js';
+export { SideNavLink } from './components/SideNavLink.js';
+export { SideNavLinkPopover } from './components/SideNavLinkPopover.js';
 export { SideNavMenu } from './components/SideNavMenu.js';
 export { SideNavMenuItem } from './components/SideNavMenuItem.js';
+export { HeaderContainer } from './components/HeaderContainer';
 export { HeaderPanel } from './components/HeaderPanel';
+export { SharkFinIcon } from './components/SharkFinIcon';